test: start separating out amp tests from stable pool tests
